refactor: 메뉴에 대한 컴포넌트 분리
